+++ src/cql-service/cqlService.executeCql.ts
@@ -7,12 +7,12 @@ import { CqlParametersConfig } from '../model/parameters';
 import { resolveParameters } from '../helpers/parametersHelper';
 import { extractLibraryVersion } from '../helpers/fileHelper';
 import { TestCase } from '../model/testCase';
-import { VersionInfo } from '../protocol';
+import { VersionInfoWithExtension } from '../commands/execute-cql-file';
 
 export interface ExecuteCqlResponse {
   results: LibraryResult[];
   logs: string[];
-  versions?: VersionInfo;
+  versions?: VersionInfoWithExtension;
 }
 
 export interface LibraryResult {